requirements.txt
jax==0.5.2
flax==0.10.5
huggingface-hub==0.30.1
python-dotenv
tiktoken==0.9.0
safetensors==0.5.3
tqdm==4.67.1
numpy==2.0.2
blobfile==3.0.0